feat: Add IM Softworks modal and Bengali font
Introduces a new modal component for "IM Softworks" with an "About" section. This modal is designed to provide information about the software.

Additionally, this commit integrates a custom Bengali font using a base64 encoded string. This font will be used for any future localization or display requirements involving Bengali text.